moving the custom code to a rescue block and removing the 2 provider because pain
