+++ lib/puppet/provider/sysctl/linux.rb
@@ -1,8 +1,36 @@
 Puppet::Type.type(:sysctl).provide(:linux) do
 
   confine  :kernel => 'linux'
-  confine  :true   => Puppet.version =~ /^2/
-  commands :sysctl => 'sysctl'
+
+  begin
+    # This code contributed by Tom Doran to fix the stdout/err stream merging
+    # problem where Puppet doesn't allow the command helper to specify 
+    # or allow handling of stderr from commands with biggish output
+    # Wrapped in a BeginRescueEnd because puppet 2.x
+
+    class CommandDefinerNoMerge < Puppet::Provider::CommandDefiner
+      def command
+        @confiner.confine :exists => @path, :for_binary => true
+        Puppet::Provider::Command.new(@name, @path, Puppet::Util, Puppet::Util::Execution, { :failonfail => true, :combine => false, :custom_environment => @custom_environment })
+      end
+    end
+ 
+    def self.has_nomerge_command(name, path, &block)
+      name = name.intern
+      command = CommandDefinerNoMerge.define(name, path, self, &block)
+  
+      @commands[name] = command.executable
+  
+      create_class_and_instance_method(name) do |*args|
+        return command.execute(*args)
+      end
+    end
+
+    has_nomerge_command(:sysctl, 'sysctl') do
+    end
+  rescue
+    commands :sysctl => 'sysctl'
+  end
 
   def exists?
     @property_hash[:ensure] == :present
